imp: ROMAlyzer: slightly changed the way the check-sum scanner works to make future support for "incremental scans" possible / added a progress bar to the scanner log
imp: using qint64 instead of int for row-counters in all DB managers

+++ checksumdbmgr.h
@@ -22,7 +22,7 @@ class CheckSumDatabaseManager : public QObject
 		uint scanTime();
 		void setScanTime(uint scan_time);
 
-		int checkSumRowCount();
+		qint64 checkSumRowCount();
 
 		QString connectionName() { return m_connectionName; }
 		QString databasePath() { return m_db.databaseName(); }
